-- Deleted var "holes" -- redesign -- new logic for CombinedObject.cs -- new logic for Data.cs -- renamed some classes/interface

+++ MainProject/MainProject/Models/RoundObject.cs
@@ -6,7 +6,7 @@
         public double Y { get; private set; }
         public double Z { get; private set; }
 
-        public Point(int x, int y, int z)
+        public Point(double x, double y, double z)
         {
             X = x;
             Y = y;
